# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Rest API document \u003cstyle\u003e.models {display: none !important}\u003c/style\u003e",
        "title": "COJ Specialist API",
        "termsOfService": "COJ",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "เข้าสู่ระบบ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "เข้าสู่ระบบ",
                "parameters": [
                    {
                        "description": "ลงชื่อเข้าสู่ระบบ",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/routes.login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SwagLogin"
                        }
                    }
                }
            }
        },
        "/auth/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "ข้อมูลตนเอง",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "ข้อมูลตนเอง",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SwagGetUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError500"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "สมัครเข้าใช้งาน",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "สมัครเข้าใช้งาน",
                "parameters": [
                    {
                        "description": "สมัครเข้าใช้งาน",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterUser"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SwagCreateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError500"
                        }
                    }
                }
            }
        },
        "/roles": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "รายการสิทธิ์การใช้งาน",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Role"
                ],
                "summary": "รายการสิทธิ์การใช้งาน",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SwagGetRoleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError500"
                        }
                    }
                }
            }
        },
        "/users": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "รายการผู้ใช้งาน",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "รายการผู้ใช้งาน",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SwagGetAllUsersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError500"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "เพิ่มผู้ใช้งาน",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "เพิ่มผู้ใช้งาน",
                "parameters": [
                    {
                        "description": "เพิ่มผู้ใช้งาน",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwagUserBodyIncludePassword"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SwagCreateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError500"
                        }
                    }
                }
            }
        },
        "/users/{slug}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "ข้อมูลผู้ใช้งาน",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "ข้อมูลผู้ใช้งาน",
                "parameters": [
                    {
                        "type": "string",
                        "description": "slug ผู้ใช้งาน",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SwagGetUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError500"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "แก้ไขผู้ใช้งาน",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "แก้ไขผู้ใช้งาน",
                "parameters": [
                    {
                        "type": "string",
                        "description": "slug ผู้ใช้งาน",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "แก้ไขผู้ใช้งาน",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SwagUserBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SwagUpdateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError500"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "ลบผู้ใช้งาน",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "ลบผู้ใช้งาน",
                "parameters": [
                    {
                        "type": "string",
                        "description": "slug ผู้ใช้งาน",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SwagDeleteBase"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError500"
                        }
                    }
                }
            }
        },
        "/users/{slug}/avatar": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "อัพโหลดรูป avatar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "อัพโหลดรูป avatar",
                "parameters": [
                    {
                        "type": "string",
                        "description": "slug ผู้ใช้งาน",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "อัพโหลดรูป avatar",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UploadAvatar"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SwagUploadAvatarResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError500"
                        }
                    }
                }
            }
        },
        "/users/{slug}/password": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "เปลี่ยนรหัสผ่าน",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "เปลี่ยนรหัสผ่าน",
                "parameters": [
                    {
                        "type": "string",
                        "description": "slug ผู้ใช้งาน",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "เปลี่ยนรหัสผ่าน",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ChangePassword"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.SwagChangePasswordResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.SwagError500"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ChangePassword": {
            "type": "object",
            "required": [
                "current_password",
                "new_password"
            ],
            "properties": {
                "current_password": {
                    "description": "รหัสผ่านปัจจุบัน",
                    "type": "string",
                    "example": "password"
                },
                "new_password": {
                    "description": "รหัสผ่านใหม่",
                    "type": "string",
                    "example": "password123"
                }
            }
        },
        "models.RegisterUser": {
            "type": "object",
            "required": [
                "email",
                "firstName",
                "lastName",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "models.SwagChangePasswordResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "ข้อมูล",
                    "type": "object"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Change Password Successfully"
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SwagCreateUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/models.SwagUser"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Created Successfully"
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SwagDeleteBase": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "ข้อมูล",
                    "type": "object"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Deleted Successfully"
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SwagError400": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "ข้อมูล",
                    "type": "object"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Bad Request"
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.SwagError404": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "ข้อมูล",
                    "type": "object"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Not Found"
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.SwagError500": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "ข้อมูล",
                    "type": "object"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Internal Server Error"
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "models.SwagGetAllUsersResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.SwagUser"
                    }
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Data Found"
                },
                "pageMeta": {
                    "type": "object",
                    "$ref": "#/definitions/models.SwagPageMeta"
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SwagGetRoleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/models.Role"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Data Found"
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SwagGetUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/models.SwagUser"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Data Found"
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SwagLogin": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/models.SwagUser"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "User logged in successfully"
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": true
                },
                "token": {
                    "description": "token",
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsI"
                }
            }
        },
...
        "models.SwagUpdateUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "$ref": "#/definitions/models.SwagUser"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Updated Successfully"
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "models.SwagUploadAvatarResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "ข้อมูล",
                    "type": "object"
                },
                "message": {
                    "description": "ข้อความตอบกลับ",
                    "type": "string",
                    "example": "Uploaded Avatar Successfully"
                },
                "success": {
                    "description": "ผลการเรียกใช้งาน",
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "models.SwagUserBody": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "รูป Avatar",
                    "type": "string",
                    "example": "user01.png"
                },
                "email": {
                    "description": "อีเมล์",
                    "type": "string",
                    "example": "user01@email.com"
                },
                "firstName": {
                    "description": "ชื่อ",
                    "type": "string",
                    "example": "john"
                },
                "lastName": {
                    "description": "นามสกุล",
                    "type": "string",
                    "example": "doe"
                },
                "slug": {
                    "description": "Slug",
                    "type": "string",
                    "example": "user01"
                },
                "username": {
                    "description": "Username",
                    "type": "string",
                    "example": "user01"
                }
            }
        },
        "models.SwagUserBodyIncludePassword": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "รูป Avatar",
                    "type": "string",
                    "example": "user01.png"
                },
                "email": {
                    "description": "อีเมล์",
                    "type": "string",
                    "example": "user01@email.com"
                },
                "firstName": {
                    "description": "ชื่อ",
                    "type": "string",
                    "example": "john"
                },
                "lastName": {
                    "description": "นามสกุล",
                    "type": "string",
                    "example": "doe"
                },
                "password": {
                    "description": "รหัสผ่าน",
                    "type": "string",
                    "example": "pass1234"
                },
                "slug": {
                    "description": "Slug",
                    "type": "string",
                    "example": "user01"
                },
                "username": {
                    "description": "Username",
                    "type": "string",
                    "example": "user01"
                }
            }
        },
        "models.UploadAvatar": {
            "type": "object",
            "required": [
                "avatar"
            ],
            "properties": {
                "avatar": {
                    "description": "รูปภาพ",
                    "type": "string",
                    "example": "avatar.png"
                }
            }
        },
        "routes.login": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}